model_training/requirements.txt
tensorflow-gpu
pandas
pillow